Added Pop Up for Interaction
When you look at something that implements the "InteractInterface",(anything that is interactable with must have this) a widget is called which tells the player to press E to interact.
Needs better style on the widget.
Needs Text needs to be made variable for when the player can change keybinds