Many changes to Plant Pot, improved A LOT
The Plant Pot now grows plants.
Empty the planter by interacting with it to plant new plants.
Interact with the plants to collect them if they've reached a certain age.

APlant (class)
Parameters:
TimePerStage; the time the plants needs to progress to it's next stage
GrowthStates; the stages the plant can run through
GrowFactor; How much the plant grows per stage. NOTE: Currently plants don't change models when growing, they only change their size.

MinHarvestableState & MaxHarvestableState; The time when a plant can be harvested. NOTE: If the plant grows over its MaxHarvestableState it can't be collected. This is meant to be for when the plant isn't harvested in time. Set MaxHarvestableState to  GrowthStages if this is not needed.

BottomStemThickness; This is solely used when spawning the plant. Set the value to the radius of the plant, don't add anything extra. It makes sure the plant doesn't spawn in something, e.g. the walls of the planter.

Functions:
GrowPlant; advances the plants grow state to the next state if it is not already in its maximum stage. Gets called after TimePerStage is reached in a cycle but can be called manually if needed.
PickupPlant; Removes the plant and adds it to the players inventory if it is in a harvestable state

APlantPot (class)
Paramters:
Width & length; Set this to the distance of the spawned planters INNER sides opposite of each other corresponding to where the width and the length is.

Plant; The class of the plant that should be grown
MainPlant; only used internally, use this to get data from the spawned plants, since those are stored in an array.

PlantedPlant; An array containing all spawned plants.

Functions:
GetRandomPlantSpawnPoint; returns a random valid point where a plant can be spawned, mainly used internally, can be used if needed

SpawnPlants; Takes in a number and spawns as many plants of the currently selected type (variable 'Plant')

DestroyAllPlants; Destroys all plants in the planter.
SetNewPlant; Set a new plant to be grown. Only changes plant type if no plant is currently in the planter (call DestroyAllPlants).
Takes optional arguments 'bSpawnPlants', defaults to false, which also spawns the plants if set to true, and 'AmountOfPlants', which defaults to 20 and sets the amount of plants to spawn when 'bSpawnPlants' is set to true.

Further changes and improvements:
Added first plant: Bamboo!
Changed the Destroyed bridge to consume 2 bamboo.

Deleted the Widget that shows what item is needed and replaced it with one that works more dynamic and can also be controlled via C++.
Made a function in it to set the Text, which tells you how much of the item you need, to easily set the text with the optional argument of showing the name of the item needed.

Changed Items to use the Item_Base C++ base class to create items instead of using a BP class that derives from the C++ class.

Added the option to make Items which are not collectible. This might be needed later on, or for items that should only be collectable after a certain event/time/etc...