More editor QoL Updates!
You can now:
Set the amount of plants to spawn in a planter.
Set the start growth state of a plant.

Derive from the ProgressibleObject Cpp class to make objects/classes that can change their state. Set the amount, name and thumbnail of the item in the settings of the new object. This handles making a pop up prompt to tell you how many of what item you are missing. Setting the name is crucial for removing the item from the inventory, so set this to the actual name of the item you want to be used up.
Set the Stages to however many stages you want to have in the object. Leave Stage at default unless you know what you are doing.