Moving Saving to MainGameInstance
Most of the saving of the PlayerCharacter  is now happening in the MainGameInstance. This approach is more dynamic and less prune to human errors.
Building on this multiple actors will be made possible to be saved
Fixed an issue where two save games were created, because of a naming issue. This resulted in the game not being able to load that file.