Added a base class for NPCs, Refractorings, code optimizations
Added a base class that all NPC type characters should derive from.
Refractored and reformatted some code in the character class
Improved the AI Task 'AttackPlayer'. This task now also gives the option to be set up in a way to make the NPC abort it's attack when it's health get's to a certain threshold WIP