Major Inventory changes
Slightly changed the players Inventory design
Added a separate Inventory widget for not player Inventories
Added Inventory trash slot.  Put items in here if you want the destroyed

Drag and drop update:
Drop an Inventory on another one to transfer their items as far as possible
Drag an item on a different inventory to add it to the first free slot. Note: If there's no free slot it will remain where it was

Widgets now fully lock player input to the widget

Added a function to the InventoryComponent to allow adding and removing items through a 1D or 2D index.

Setting the Inventory size now works as intended and doesn't cut of rows if the wrong numbers are selected

Bug Note: When changing the constructor of the CppPlayerCharacter or adding him to the level the relative location of the mini map component is set to the wrong number. Reset the values to fix this after playing