First usage implementation of the Radial menu;
The planter can now be planted by using the radial menu.
Also added a new interact interface function, which allows passing the players selected item to make things like planting possible, by passing the selected plant. This function requires a bool return type, to determine whether the item was used up (true) or not (false).
Item_Struct now requires to have it's UClass set. This should be set to whatever class the struct is in