Almost finished T0 of RockSheep AI
Added a custom Controller for the RockSheep.
The Crystal on the back of the RockSheep now respawns after x seconds.
The Respawn/Destruction of the Crystal is now bound to a delegate.
Improved the stability of all TaskNodes (HarvestResource is still a placeholder and doesn't do anything).
The SpawnerVolume now supports the spawning of AI. NOTE: The current way the AI is setup doesn't allow them to differentiate between friend and foe, so AI spawned through the SpawnerVolume will attack each other. This can be worked around by manually placing them.
Improved the ease of use for the item collectability system.