Added a Spawner Platform
Added a Platform on which the player can have things built for money.
Will add support for multiple items and multiple options for paying (either with resources and/or credits).
Added a special class for items which are to be built on this Platform.
The player can now kill the RockSheep by interacting with it (this is mainly for testing purposes).
The RockSheep can now be tamed, which makes it not attack the player anymore and not run away from them.

Added two more functions to the InventoryComponent.
RemoveItems; takes in an array of item (names)to remove
AddItems; takes in an array of items to add